The main change of this version is the implant can be generated.
The algorithms used are implictmodeler and boolean operation.
Firstly we manually draw a closed loop on the surface of the model, then
we select all the cells which contain the every point in the loop, and we
delete all the cell from the model. with the vtkconnectpolydatafilter, we
can extract the surface inside the loop. then we use the surface to
generate a distance filed which can be contoured with a specified value
into a model. then we do boolean operation between the origin model and
the contoured model which will give us the final model as the implant. But
the boolean operation can cause some hole on the implant which is
unavoidable. I think we can use some tool to fill the hole, not a very big
deal. what mater is the speed of doing boolean operation.